Implement getting existing user records
With this commit I have fully implemented the functionality that allows
retrieving user details from file.
Currently, the user files have to be generated manually. Functionality to
save user data to file will be coming next
The files are stored in the following format:
    - They are named as: <account_number>.txt e.g. 501.txt
    - Each data is stored in it's own line as per the following order:
        account_number
        first_name
        last_name
        balance
    - The files are stored in a folder in the src folder.
        This folder is called accounts.

The above requirements need to be fulfilled in order for the program to work
as intended.
Due to time constrictions the code is crowded and with little documentation
at the moment. That will be dealt with in future commits.

Signed-off-by: Vickz84259 <vickz84259@gmail.com>